Review on Add flanges and Add valves classes
