+++ pulse/interface/user_input/project/call_double_confirmation.py
@@ -42,18 +42,19 @@ class CallDoubleConfirmationInput(QDialog):
         self._doNotRun = True
 
     def _define_qt_variables(self):
-        self.label_message = self.findChild(QLabel, 'label_message')
-        self.label_title = self.findChild(QLabel, 'label_title')
-        self.pushButton_rightButton = self.findChild(QPushButton, 'pushButton_rightButton')
-        self.pushButton_leftButton = self.findChild(QPushButton, 'pushButton_leftButton')
-    
+        # QLabel
+        self.label_message : QLabel
+        self.label_title : QLabel
+        # QPushButton
+        self.pushButton_rightButton : QPushButton
+        self.pushButton_leftButton : QPushButton
+
     def _create_actions(self):
         self.pushButton_rightButton.clicked.connect(self.confirm_action)
         self.pushButton_leftButton.clicked.connect(self.force_to_close)
 
     def _configure_buttons(self):
-
-        if self.buttons_config != {}:
+        if self.buttons_config:
             if "left_button_label" in self.buttons_config.keys():
                 self.pushButton_leftButton.setText(self.buttons_config["left_button_label"])
             if "right_button_label" in self.buttons_config.keys():
